various fixes

+++ browser/containers/Application.js
@@ -45,6 +45,8 @@ class ApplicationContainer extends React.Component {
 
     this.handleChangeInTerm = this.handleChangeInTerm.bind(this)
     
+    this.validateFields = this.validateFields.bind(this)
+    this.generateErrors = this.generateErrors.bind(this)
     this.handleSave = this.handleSave.bind(this)
     this.handleSubmit = this.handleSubmit.bind(this)
     this.handleDelete = this.handleDelete.bind(this)
@@ -180,8 +182,27 @@ class ApplicationContainer extends React.Component {
     })
   }
 
+  validateFields() {
+    let errors = {}
+    if (this.state.customer && this.state.customer.phone) errors.phone = (this.state.customer.phone.search(/\d\d\d-\d\d\d-\d\d\d\d|\(\d\d\d\)\s\d\d\d-\d\d\d\d/) === -1) ? true : false
+    if (this.state.customer && this.state.customer.email) errors.email = (this.state.customer.email.search(/\w+@\w+\.\w+/) === -1) ? true : false
+    return errors
+  }
+
+  generateErrors(errors) {
+    let fields = Object.keys(errors).filter(n => errors[n])
+
+    return (e) => {
+      e.preventDefault()
+      this.props.throwAlert('red', `Problem with the ${fields.join(', ')} field${(fields.length > 1) ? 's' : ''}`)
+    }
+
+  }
+
   handleSave(e) {
     e.preventDefault()
+
+    
 
     let custArr = (this.state.customerCreate) ? [this.state.customer, {id: 'new'}] : [this.state.customer]
     this.props.saveAppThunk(this.props.token, [this.state, {amount: checkFor$(this.state.amount)}], custArr)
@@ -238,7 +259,7 @@ class ApplicationContainer extends React.Component {
 
   handleDelete(e) {
     e.preventDefault()
-    this.props.deleteAppThunk(this.props.token, this.state.id, () => {this.props.history.push('/applications')})
+    if (confirm('Are you sure you wish to delete this application?')) this.props.deleteAppThunk(this.props.token, this.state.id, () => {this.props.history.push('/applications')})
   }
 
   handleCheckbox(e) {
@@ -303,7 +324,10 @@ class ApplicationContainer extends React.Component {
       // this.props.throwAlert('green', 'Success')
       this.setState({adminMode: false})
     })
-    .catch(err => console.error(err))
+    .catch(err => {
+      console.error(err)
+      this.props.throwAlert('red', 'Something went wrong')
+    })
 
   }
 
@@ -320,11 +344,14 @@ class ApplicationContainer extends React.Component {
       // this.props.throwAlert('green', 'Success')
       this.setState({adminMode: 'notify'})
     })
-    .catch(err => console.error(err))
+    .catch(err => {
+      console.error(err)
+      this.props.throwAlert('red', 'Something went wrong')
+    })
   }
 
   handleAdminMode(e) {
-    console.log('trigger', e.target.id)
+    // console.log('trigger', e.target.id)
     if (e.target.id === 'cancel-button') {
       this.setState({adminMode: false})
     } else if (e.target.id === 'submit-button' || e.target.id === 'app-button'){
@@ -404,10 +431,14 @@ class ApplicationContainer extends React.Component {
         else return term
       } else return term
     }
+    let errors = this.validateFields()
+    let disabled = Object.keys(errors).some(n => errors[n])
+
     return(
       <div>
         <EditApplication
           values={this.state}
+          errors={errors}
           iAmAuthor={(this.props.user) ? this.props.user.email === this.state.rep.email : false}
           admin={(this.props.user) ? this.props.user.level === 'Admin' : false}
           customers={(this.props.customers) ? this.props.customers.map(customer => customer.name) : null}
@@ -423,8 +454,8 @@ class ApplicationContainer extends React.Component {
           handleChangeInMachine={this.handleChangeInMachine}
           handleRemoveMachine={this.handleRemoveMachine}
           handleChangeInCustomer={this.handleChangeInCustomer}
-          handleSave={this.handleSave}
-          handleSubmit={this.handleSubmit}
+          handleSave={(disabled) ? this.generateErrors(errors) : this.handleSave}
+          handleSubmit={(disabled) ? this.generateErrors(errors) : this.handleSubmit}
           handleDelete={this.handleDelete}
           handleChangeCustomer={this.handleChangeCustomer}
           handleCheckbox={this.handleCheckbox}
